custom-elements/registering-custom-elements/definition-construction-algorithm-test-004.html

<!DOCTYPE html>
<html>
<head>
<title>Check duplicate definition</title>
<meta name="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<meta name="assert" content="If there already exists a definition with the same TYPE, set ERROR to DuplicateDefinition and stop">
<link rel="help" href="http://www.w3.org/TR/custom-elements/#registering-custom-elements">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src='../testcommon.js'></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script type="text/javascript">

test(function() {
    var doc = newHTMLDocument();
    var name = 'x-a-1';
    doc.register(name);
    //FIXME Should the type of the exception be specified here?
    assert_throws(null, function() {
        doc.register(name);    
    }, 'Exception should be thrown if definition with the same type already exists');
}, 'Check duplicate definition');


test(function() {
    var doc = newHTMLDocument();
    var name = 'x-a-2';
    doc.register(name);
    
    HTML5_ELEMENTS.forEach(function (element) {
        //FIXME Should the type of the exception be specified here?
        assert_throws(null, function() {
            doc.register(name, {
                prototype: Object.create(bj.constructor.prototype),
                extends: element       
            });    
        }, 'Exception should be thrown if definition with the same type already exists');        
    });
}, 'Check duplicate definition. Specify constructor');


test(function() {
    var doc = newHTMLDocument();    
    var name = 'x-a-3';
    doc.register(name, {
        prototype: Object.create(HTMLAnchorElement.prototype),
        extends: 'a'
    });
    
    HTML5_ELEMENTS.forEach(function(element) {
        //FIXME Should the type of the exception be specified here?
        assert_throws(null, function() {
            doc.register(name, {
                prototype: Object.create(bj.constructor.prototype),
                extends: element       
            });    
        }, 'Exception should be thrown if definition with the same type already exists');        
    });
}, 'Check duplicate definition. Test different prototypes and extends');

</script>
</body>
</html>
